fix: binary paths on env file
